+++ src/process.h
@@ -2,8 +2,8 @@
 #define __PROCESS_H__
 #include <stdio.h>
 #include "neutron.h"
-void getASE(float *accmeans,unsigned meshes, unsigned nbat, unsigned ubat, float ref, float* ASE);
+void getASE(double *accmeans,unsigned meshes, unsigned nbat, unsigned ubat, double ref, double* ASE);
 void cnt2flux(MemStruct HostMem, unsigned numhis, float dx, unsigned meshes, unsigned nbat);
-float autok(float *batmeans, unsigned n, unsigned k, unsigned meshes, unsigned im);
-void getCOR(float *batmeans, unsigned meshes, unsigned nbat, unsigned ubat, unsigned upto,float *COR);
+double autok(double *batmeans, unsigned n, unsigned k, unsigned meshes, unsigned im);
+void getCOR(double *batmeans, unsigned meshes, unsigned nbat, unsigned ubat, unsigned upto,double *COR);
 #endif
